+++ CHANGELOG.md
@@ -20,6 +20,7 @@
     - Removed the distinction between "visible" sets and "selected" (i.e. checked) sets.
 - Upgrade deck.gl to 8.3 and viv to 0.5
 - Fix build issues with nebula.
+- Store updates to the grid layout (removing / moving / resizing) in the view config zustand store.
 
 ## [1.0.0](https://www.npmjs.com/package/vitessce/v/1.0.0) - 2020-09-2